Refactor and prune dependencies
