+++ pyserini/output_writer.py
@@ -21,7 +21,7 @@ from abc import ABC, abstractmethod
 from enum import Enum, unique
 from typing import List
 
-from pyserini.search import JScoredDoc
+from pyserini.search.lucene import JScoredDoc
 
 
 @unique
