+++ pyserini/search/lucene/__init__.py
@@ -14,17 +14,26 @@
 # limitations under the License.
 #
 
-from ._geo_searcher import LuceneGeoSearcher
-from ._impact_searcher import JScoredDoc, LuceneImpactSearcher, SlimSearcher
-from ._searcher import JScoredDoc, LuceneSimilarities, LuceneFusionSearcher, LuceneSearcher
-from ._hnsw_searcher import LuceneHnswDenseSearcher, LuceneFlatDenseSearcher
+# from ._geo_searcher import LuceneGeoSearcher
+# from ._impact_searcher import JScoredDoc, LuceneImpactSearcher, SlimSearcher
+#from ._searcher import JScoredDoc
+# from ._hnsw_searcher import LuceneHnswDenseSearcher, LuceneFlatDenseSearcher
+#
+# __all__ = ['JScoredDoc',
+#            'LuceneFusionSearcher',
+#            'LuceneGeoSearcher',
+#            'LuceneImpactSearcher',
+#            'LuceneSearcher',
+#            'LuceneHnswDenseSearcher',
+#            'LuceneFlatDenseSearcher',
+#            'SlimSearcher',
+#            'LuceneSimilarities']
 
-__all__ = ['JScoredDoc',
-           'LuceneFusionSearcher',
-           'LuceneGeoSearcher',
-           'LuceneImpactSearcher',
-           'LuceneSearcher',
-           'LuceneHnswDenseSearcher',
-           'LuceneFlatDenseSearcher',
-           'SlimSearcher',
-           'LuceneSimilarities']
+from pyserini.pyclass import autoclass
+
+JQuery = autoclass('org.apache.lucene.search.Query')
+JScoredDoc = autoclass('io.anserini.search.ScoredDoc')
+JQueryGenerator = autoclass('io.anserini.search.query.QueryGenerator')
+JBagOfWordsQueryGenerator = autoclass('io.anserini.search.query.BagOfWordsQueryGenerator')
+JDisjunctionMaxQueryGenerator = autoclass('io.anserini.search.query.DisjunctionMaxQueryGenerator')
+JCovid19QueryGenerator = autoclass('io.anserini.search.query.Covid19QueryGenerator')
