+++ requirements.txt
@@ -1,19 +1,20 @@
 Cython>=0.29.21
 numpy>=1.18.1
 pandas>=1.4.0
-pyjnius>=1.4.0
+pyjnius>=1.6.0
 scikit-learn>=0.22.1
 scipy>=1.4.1
 tqdm
 transformers>=4.6.0
-sentencepiece>=0.1.95
-nmslib>=2.0.6
+torch>=2.2.0
 onnxruntime>=1.8.1
-lightgbm>=3.3.2
-spacy>=3.2.1
 pyyaml
 openai>=1.0.0
-tiktoken>=0.4.0
-pyarrow>=15.0.0
-pillow>=10.2.0
-pybind11>=2.11.0
+#sentencepiece>=0.1.95
+#nmslib>=2.0.6
+#lightgbm>=3.3.2
+#spacy>=3.2.1
+#tiktoken>=0.4.0
+#pyarrow>=15.0.0
+#pillow>=10.2.0
+#pybind11>=2.11.0
